Separate files for IAuthenticationContext and stuff
